requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
altgraph==0.17.2
appdirs==1.4.4
asgiref==3.5.0
astroid==2.4.2
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
branca==0.3.1
cattrs==1.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
dj-database-url==0.5.0
Django==3.2.12
django-bootstrap4==22.1
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-fontawesome==1.0
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
entrypoints==0.3
fastf1==2.1
findspark==1.4.2
folium==0.12.1.post1
fonttools==4.32.0
frozenlist==1.3.0
future==0.18.2
fuzzywuzzy==0.18.0
geographiclib==1.52
geoip2==4.5.0
gunicorn==20.1.0
idna==3.3
ipykernel==6.6.0
ipython==7.30.1
isort==5.7.0
jedi==0.18.1
Jinja2==3.1.1
jupyter-client==7.1.0
jupyter-core==4.9.1
kiwisolver==1.4.2
lazy-object-proxy==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
maxminddb==2.2.0
mccabe==0.6.1
metakernel==0.28.2
mongoengine==0.24.1
multidict==6.0.2
nest-asyncio==1.5.4
numpy==1.22.3
packaging==21.3
pandas==1.4.2
parso==0.8.3
pefile==2021.9.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pycodestyle==2.8.0
Pygments==2.10.0
pyinstaller-hooks-contrib==2022.3
pylint==2.6.0
pymongo==3.12.3
pyparsing==3.0.8
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
python-snappy==0.6.1
pytz==2022.1
pywin32==303
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-cache==0.5.2
scipy==1.8.0
seaborn==0.11.2
signalr-client-aio==0.0.1.6.2
six==1.15.0
soupsieve==2.3.2.post1
spylon==0.3.0
spylon-kernel==0.4.1
sqlparse==0.2.4
thefuzz==0.19.0
timple==0.1.5
toml==0.10.2
tornado==6.1
traitlets==5.1.1
tzdata==2022.1
url-normalize==1.4.3
urllib3==1.26.9
wcwidth==0.2.5
websockets==10.3
whitenoise==6.0.0
winkerberos==0.8.0
wrapt==1.12.1
yarl==1.7.2
